+++ src/header_authorization_jwt_token_.js
@@ -1,7 +1,10 @@
-import type { nullish } from '@ctx-core/function'
-export function header_authorization_jwt_token_(authorization:string|nullish):string|undefined {
+/**
+ * @param {string|nullish}authorization
+ * @return {string|null}
+ */
+export function header_authorization_jwt_token_(authorization) {
 	const authorization_a = authorization?.split(/^Bearer */)
-	return authorization_a ? authorization_a[1] : undefined
+	return authorization_a ? authorization_a[1] : null
 }
 export {
 	header_authorization_jwt_token_ as _header_authorization_jwt_token,