+++ src/validate_current_jwt.d.ts
@@ -1,0 +1,3 @@
+import type { nullish } from '@ctx-core/function'
+export declare function validate_current_jwt(jwt_token:string|nullish):void
+export { validate_current_jwt as validate__current__jwt }